Add files via upload
This is the modified version of the first code. I have added the Bootstrap framework (last two sections) and the code is supposed to plot the error, bias and variance of the training and test data under OLS, Ridge and Lasso regression, as a function of the regularisation paramaters lamda. The problem is that it's too hefty got my PC, so if someone has access to a more powerful system, please run the code on that and lemme know the output please :)