Add files via upload
This is the raw version of Part (c) Logistic regression. I haven't debugged a big part of it yet, but I will do it in some time by running it on Iftikhar's PC. All functionalities have been added except cross-validation, which is tomorrow (or maybe tonight's task). If you can run it on your PC, I'd be glad to get your input, else I'll do it in a while :)